Keep __main__ and cli separate
